pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pivot"
version = "0.1.0"
description = "Continuous localization pipeline for GitHub-hosted technical documentation."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Pivot Contributors" }
]
keywords = ["localization", "documentation", "automation"]
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12.3",
  "pydantic>=2.7.1",
  "GitPython>=3.1.43",
  "markdown-it-py>=3.0.0",
  "ruamel.yaml>=0.18.0",
  "httpx>=0.27.0",
  "rich>=13.7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.4",
  "mypy>=1.9.0"
]

[project.scripts]
pivot = "pivot.cli:app"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "W",
  "UP"
]
ignore = [
  "B008"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_reexport = true
show_error_codes = true
exclude = [
  "tests/"
]